+++ packages/flutter_migrate/test/status_test.dart
@@ -136,11 +136,7 @@ All conflicts resolved. Review changes above and apply the migration with:
     conflictFile.createSync(recursive: true);
     conflictFile.writeAsStringSync('''
 line1
-<<<<<<< /conflcit/conflict.file
 line2
-=======
-linetwo
->>>>>>> /var/folders/md/gm0zgfcj07vcsj6jkh_mp_wh00ff02/T/flutter_tools.4Xdep8/generatedTargetTemplatetlN44S/conflict/conflict.file
 line3
 ''', flush: true);
     final File conflictFileOriginal =