+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/google_maps_controller_test.dart
@@ -16,6 +16,11 @@ import 'package:mockito/mockito.dart';
 
 import 'google_maps_controller_test.mocks.dart';
 
+// This value is used when comparing long~num, like
+// LatLng values.
+const double _acceptableDelta = 0.0000000001;
+const String _kCloudMapId = '000000000000000'; // Dummy map ID.
+
 @GenerateMocks(<Type>[], customMocks: <MockSpec<dynamic>>[
   MockSpec<CirclesController>(onMissingStub: OnMissingStub.returnDefault),
   MockSpec<PolygonsController>(onMissingStub: OnMissingStub.returnDefault),
@@ -383,6 +388,7 @@ void main() {
               mapConfiguration: const MapConfiguration(
             mapType: MapType.satellite,
             zoomControlsEnabled: true,
+            cloudMapId: _kCloudMapId,
           ));
           controller.debugSetOverrides(
               createMap: (_, gmaps.MapOptions options) {
@@ -395,6 +401,7 @@ void main() {
           expect(capturedOptions, isNotNull);
           expect(capturedOptions!.mapTypeId, gmaps.MapTypeId.SATELLITE);
           expect(capturedOptions!.zoomControl, true);
+          expect(capturedOptions!.mapId, _kCloudMapId);
           expect(capturedOptions!.gestureHandling, 'auto',
               reason:
                   'by default the map handles zoom/pan gestures internally');