+++ packages/google_maps_flutter/google_maps_flutter_ios/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.3.0
+
+* Adds implementation for `cloudMapId` parameter to support cloud-based maps styling.
+
 ## 2.2.1
 
 * Clarifies explanation of endorsement in README.
@@ -10,7 +14,6 @@
 ## 2.1.14
 
 * Updates links for the merge of flutter/plugins into flutter/packages.
-* Updates minimum Flutter version to 3.0.
 
 ## 2.1.13